+++ kernel/src/main/java/org/kframework/krun/tools/Debugger.java
@@ -41,7 +41,6 @@ import org.kframework.utils.inject.Main;
 import com.beust.jcommander.JCommander;
 import com.beust.jcommander.ParameterException;
 import com.google.inject.Inject;
-import com.google.inject.Provider;
 
 public interface Debugger {
 
@@ -124,8 +123,7 @@ public interface Debugger {
 
     public static class Tool implements Transformation<Void, Void> {
 
-        private final KExceptionManager kem;
-        private final Provider<Term> initialConfiguration;
+        private final Term initialConfiguration;
         private final KompileOptions kompileOptions;
         private final BinaryLoader loader;
         @InjectGeneric private Transformation<KRunState, String> statePrinter;
@@ -139,15 +137,13 @@ public interface Debugger {
 
         @Inject
         Tool(
-                KExceptionManager kem,
-                @Main Provider<Term> initialConfiguration,
+                @Main Term initialConfiguration,
                 @Main KompileOptions kompileOptions,
                 BinaryLoader loader,
                 @Main Debugger debugger,
                 @Main Context context,
                 @Main FileUtil files,
                 @Main(Concrete.class) Definition definition) {
-            this.kem = kem;
             this.initialConfiguration = initialConfiguration;
             this.kompileOptions = kompileOptions;
             this.loader = loader;
@@ -158,8 +154,7 @@ public interface Debugger {
         }
 
         Tool(
-                KExceptionManager kem,
-                @Main Provider<Term> initialConfiguration,
+                @Main Term initialConfiguration,
                 @Main KompileOptions kompileOptions,
                 BinaryLoader loader,
                 Transformation<KRunState, String> statePrinter,
@@ -170,7 +165,7 @@ public interface Debugger {
                 @Main Context context,
                 @Main FileUtil files,
                 @Main(Concrete.class) Definition definition) {
-            this(kem, initialConfiguration, kompileOptions, loader, debugger, context, files, definition);
+            this(initialConfiguration, kompileOptions, loader, debugger, context, files, definition);
             this.statePrinter = statePrinter;
             this.searchPrinter = searchPrinter;
             this.graphPrinter = graphPrinter;
@@ -205,7 +200,7 @@ public interface Debugger {
             reader.addCompletor(new MultiCompletor(completors));
 
             try {
-                debugger.start(initialConfiguration.get());
+                debugger.start(initialConfiguration);
                 System.out.println("After running one step of execution the result is:\n");
                 System.out.println(statePrinter.run(debugger.getState(debugger.getCurrentState()), a));
             } catch (UnsupportedBackendOptionException e) {