+++ kernel/src/main/java/org/kframework/krun/tools/Executor.java
@@ -33,6 +33,8 @@ import org.kframework.utils.errorsystem.KExceptionManager;
 import org.kframework.utils.errorsystem.ParseFailedException;
 import org.kframework.utils.inject.Main;
 
+import com.google.inject.Inject;
+
 import java.util.HashSet;
 import java.util.Set;
 
@@ -61,12 +63,13 @@ public interface Executor {
     @param cfg The term to begin the search at
     @param compilationInfo the object used to kompile the search pattern, which contains
     metadata used to pretty-print results
+    @param computeGraph whether to compute the search graph
     @exception KRunExecutionException Thrown if the backend fails to successfully perform the
     search
     @return An object containing both metadata about krun's execution, and information about
     the results of the search
     */
-    public abstract SearchResults search(Integer bound, Integer depth, SearchType searchType, Rule pattern, Term cfg, RuleCompilerSteps compilationInfo) throws KRunExecutionException;
+    public abstract SearchResults search(Integer bound, Integer depth, SearchType searchType, Rule pattern, Term cfg, RuleCompilerSteps compilationInfo, boolean computeGraph) throws KRunExecutionException;
 
     /**
     Execute a term in normal-execution mode for a specified number of steps
@@ -89,7 +92,7 @@ public interface Executor {
 
         public static final String EXIT_CODE = "exitCode";
         private final KRunOptions options;
-        private final Provider<Term> initialConfiguration;
+        private final Term initialConfiguration;
         private final Context context;
         private final Stopwatch sw;
         private final KExceptionManager kem;
@@ -99,7 +102,7 @@ public interface Executor {
         @Inject
         Tool(
                 KRunOptions options,
-                @Main Provider<Term> initialConfiguration,
+                @Main Term initialConfiguration,
                 Stopwatch sw,
                 @Main Context context,
                 KExceptionManager kem,
@@ -153,7 +156,7 @@ public interface Executor {
                         options.depth,
                         options.searchType(),
                         searchPattern.patternRule,
-                        initialConfiguration.get(), searchPattern.steps);
+                        initialConfiguration, searchPattern.steps, false);
 
             sw.printIntermediate("Search total");
             return result;
@@ -162,10 +165,10 @@ public interface Executor {
         public KRunResult execute(Attributes a) throws ParseFailedException, KRunExecutionException {
             KRunState result;
             if (options.depth != null) {
-                result = executor.step(initialConfiguration.get(), options.depth, false).getFinalState();
+                result = executor.step(initialConfiguration, options.depth, false).getFinalState();
                 sw.printIntermediate("Bounded execution total");
             } else {
-                result = executor.run(initialConfiguration.get(), false).getFinalState();
+                result = executor.run(initialConfiguration, false).getFinalState();
                 sw.printIntermediate("Normal execution total");
             }
             ASTNode pattern = pattern(options.pattern);
@@ -176,7 +179,7 @@ public interface Executor {
             if (pattern != null && !options.search()) {
                 SearchPattern searchPattern = new SearchPattern(pattern);
                 Term res = result.getRawResult();
-                return executor.search(1, 1, SearchType.FINAL, searchPattern.patternRule, res, searchPattern.steps);
+                return executor.search(1, 1, SearchType.FINAL, searchPattern.patternRule, res, searchPattern.steps, false);
             }
             return result;
         }
@@ -184,7 +187,7 @@ public interface Executor {
         private int getExitCode(Term res) throws KRunExecutionException {
             ASTNode exitCodePattern = pattern(options.exitCodePattern);
             SearchPattern searchPattern = new SearchPattern(exitCodePattern);
-            SearchResults results = executor.search(1, 1, SearchType.FINAL, searchPattern.patternRule, res, searchPattern.steps);
+            SearchResults results = executor.search(1, 1, SearchType.FINAL, searchPattern.patternRule, res, searchPattern.steps, false);
             if (results.getSolutions().size() != 1) {
                 kem.registerCriticalWarning("Found " + results.getSolutions().size() + " solutions to exit code pattern. Returning 112.");
                 return 112;