+++ kernel/src/main/java/org/kframework/ktest/CmdArgs/KTestOptions.java
@@ -9,6 +9,7 @@ import org.kframework.ktest.KTestStep;
 import org.kframework.ktest.StringMatcher;
 import org.kframework.main.GlobalOptions;
 import org.kframework.utils.file.FileUtil;
+import org.kframework.utils.inject.RequestScoped;
 import org.kframework.utils.options.BaseEnumConverter;
 import org.kframework.utils.options.EnumSetConverter;
 import org.kframework.utils.options.OnOffConverter;
@@ -18,8 +19,6 @@ import org.kframework.utils.options.StringListConverter;
 import com.beust.jcommander.Parameter;
 import com.beust.jcommander.ParameterException;
 import com.beust.jcommander.ParametersDelegate;
-import com.google.inject.Inject;
-
 import java.awt.Color;
 import java.io.File;
 import java.util.ArrayList;
@@ -36,13 +35,8 @@ import java.util.Set;
  *   - Only one unnamed argument is passed
  *   - File extension is either .k or .xml
  */
+@RequestScoped
 public class KTestOptions {
-
-    public KTestOptions() {}
-
-    //TODO(dwightguth): remove in Guice 4.0
-    @Inject
-    public KTestOptions(Void v) {}
 
     public static final class KTestStepSetConverter extends EnumSetConverter<KTestStep, KTestStepConverter> {
 
@@ -195,6 +189,8 @@ public class KTestOptions {
     private boolean debug = false;
 
     public final long start = System.currentTimeMillis();
+
+    public KTestOptions() {}
 
     /**
      * Copy constructor.