+++ kernel/src/main/java/org/kframework/main/Main.java
@@ -3,7 +3,6 @@ package org.kframework.main;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.PrintStream;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -11,24 +10,30 @@ import java.util.Map;
 import java.util.ServiceLoader;
 
 import org.fusesource.jansi.AnsiConsole;
-import org.fusesource.jansi.AnsiOutputStream;
 import org.kframework.kast.KastFrontEnd;
 import org.kframework.kdoc.KDocFrontEnd;
 import org.kframework.kompile.KompileFrontEnd;
 import org.kframework.krun.KRunFrontEnd;
+import org.kframework.kserver.KServerFrontEnd;
 import org.kframework.ktest.KTestFrontEnd;
 import org.kframework.utils.errorsystem.KExceptionManager;
 import org.kframework.utils.errorsystem.KExceptionManager.KEMException;
-import org.kframework.utils.file.FileSystemModule;
-import org.kframework.utils.file.TTYInfo;
+import org.kframework.utils.file.Environment;
+import org.kframework.utils.file.WorkingDir;
+import org.kframework.utils.inject.Options;
+import org.kframework.utils.inject.SimpleScope;
 
 import com.google.inject.Guice;
+import com.google.inject.Inject;
 import com.google.inject.Injector;
+import com.google.inject.Key;
 import com.google.inject.Module;
+import com.google.inject.Provider;
 import com.google.inject.ProvisionException;
+import com.google.inject.TypeLiteral;
+import com.google.inject.name.Named;
 import com.google.inject.spi.Message;
 import com.martiansoftware.nailgun.NGContext;
-import com.martiansoftware.nailgun.ThreadLocalPrintStream;
 
 public class Main {
 
@@ -43,8 +48,8 @@ public class Main {
         if (args.length >= 1) {
 
             String[] args2 = Arrays.copyOfRange(args, 1, args.length);
-            Injector injector = getInjector(args[0], args2);
-            int result = runApplication(injector);
+            Injector injector = getInjector(args[0]);
+            int result = injector.getInstance(Main.class).runApplication(args[0], args2, new File("."), System.getenv());
             AnsiConsole.systemUninstall();
             System.exit(result);
         }
@@ -60,40 +65,49 @@ public class Main {
 
     public static void nailMain(NGContext context) {
         isNailgun = true;
-        ThreadLocalPrintStream system_out = (ThreadLocalPrintStream) System.out;
-        ThreadLocalPrintStream system_err = (ThreadLocalPrintStream) System.err;
         if (context.getArgs().length >= 1) {
-
             String[] args2 = Arrays.copyOfRange(context.getArgs(), 1, context.getArgs().length);
-            Injector injector = getInjector(new File(context.getWorkingDirectory()), (Map)context.getEnv(), context.getArgs()[0], args2);
-
-            TTYInfo tty = injector.getInstance(TTYInfo.class);
-            if (tty.stdout) {
-                system_out.init(new PrintStream(AnsiConsole.wrapOutputStream(system_out.getPrintStream())));
-            } else {
-                system_out.init(new PrintStream(new AnsiOutputStream(system_out.getPrintStream())));
-            }
-            if (tty.stderr) {
-                system_err.init(new PrintStream(AnsiConsole.wrapOutputStream(system_err.getPrintStream())));
-            } else {
-                system_err.init(new PrintStream(new AnsiOutputStream(system_err.getPrintStream())));
-            }
-
-            int result = runApplication(injector);
-            System.out.flush();
-            System.err.flush();
+            KServerFrontEnd kserver = KServerFrontEnd.instance();
+            int result = kserver.run(context.getArgs()[0], args2, new File(context.getWorkingDirectory()), (Map) context.getEnv());
             System.exit(result);
             return;
         }
         invalidJarArguments();
     }
 
-    public static int runApplication(Injector injector) {
-        KExceptionManager kem = injector.getInstance(KExceptionManager.class);
+    private final Provider<KExceptionManager> kem;
+    private final Provider<FrontEnd> frontEnd;
+    private final SimpleScope requestScope;
 
+    @Inject
+    public Main(
+            Provider<KExceptionManager> kem,
+            Provider<FrontEnd> frontEnd,
+            @Named("requestScope") SimpleScope requestScope) {
+        this.kem = kem;
+        this.frontEnd = frontEnd;
+        this.requestScope = requestScope;
+    }
+
+    public SimpleScope getRequestScope() {
+        return requestScope;
+    }
+
+    public int runApplication(String tool, String[] args, File workingDir, Map<String, String> env) {
+        try {
+            requestScope.enter();
+            seedInjector(requestScope, tool, args, workingDir, env);
+            return runApplication();
+        } finally {
+            requestScope.exit();
+        }
+    }
+
+    public int runApplication() {
+        KExceptionManager kem = this.kem.get();
         kem.installForUncaughtExceptions();
         try {
-            int retval = injector.getInstance(FrontEnd.class).main();
+            int retval = frontEnd.get().main();
             return retval;
         } catch (ProvisionException e) {
             for (Message m : e.getErrorMessages()) {
@@ -109,7 +123,14 @@ public class Main {
         }
     }
 
-    public static Injector getInjector(File workingDir, Map<String, String> env, String tool, String[] args2) {
+    public static void seedInjector(SimpleScope scope, String tool, String[] args, File workingDir,
+            Map<String, String> env) {
+        scope.seed(Key.get(File.class, WorkingDir.class), workingDir);
+        scope.seed(Key.get(new TypeLiteral<Map<String, String>>() {}, Environment.class), env);
+        scope.seed(Key.get(String[].class, Options.class), args);
+    }
+
+    public static Injector getInjector(String tool) {
         ServiceLoader<KModule> kLoader = ServiceLoader.load(KModule.class);
         List<KModule> kModules = new ArrayList<>();
         for (KModule m : kLoader) {
@@ -119,8 +140,11 @@ public class Main {
         List<Module> modules = new ArrayList<>();
 
             switch (tool) {
+                case "-kserver":
+                    modules.addAll(KServerFrontEnd.getModules());
+                    break;
                 case "-kompile":
-                    modules.addAll(KompileFrontEnd.getModules(args2));
+                    modules.addAll(KompileFrontEnd.getModules());
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getKompileModules();
                         if (ms != null) {
@@ -129,7 +153,7 @@ public class Main {
                     }
                     break;
                 case "-ktest":
-                    modules.addAll(KTestFrontEnd.getModules(args2));
+                    modules.addAll(KTestFrontEnd.getModules());
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getKTestModules();
                         if (ms != null) {
@@ -138,7 +162,7 @@ public class Main {
                     }
                     break;
                 case "-kdoc":
-                    modules.addAll(KDocFrontEnd.getModules(args2));
+                    modules.addAll(KDocFrontEnd.getModules());
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getKDocModules();
                         if (ms != null) {
@@ -147,7 +171,7 @@ public class Main {
                     }
                     break;
                 case "-kast":
-                    modules.addAll(KastFrontEnd.getModules(args2));
+                    modules.addAll(KastFrontEnd.getModules());
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getKastModules();
                         if (ms != null) {
@@ -157,7 +181,7 @@ public class Main {
                     break;
                 case "-krun":
                     List<Module> definitionSpecificModules = new ArrayList<>();
-                    definitionSpecificModules.addAll(KRunFrontEnd.getDefinitionSpecificModules(args2));
+                    definitionSpecificModules.addAll(KRunFrontEnd.getDefinitionSpecificModules());
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getDefinitionSpecificKRunModules();
                         if (ms != null) {
@@ -165,7 +189,7 @@ public class Main {
                         }
                     }
 
-                    modules.addAll(KRunFrontEnd.getModules(args2, definitionSpecificModules));
+                    modules.addAll(KRunFrontEnd.getModules(definitionSpecificModules));
                     for (KModule kModule : kModules) {
                         List<Module> ms = kModule.getKRunModules(definitionSpecificModules);
                         if (ms != null) {
@@ -174,7 +198,7 @@ public class Main {
                     }
                     break;
                 case "-kpp":
-                    modules = KppFrontEnd.getModules(args2);
+                    modules = KppFrontEnd.getModules();
                     break;
                 default:
                     invalidJarArguments();
@@ -184,7 +208,6 @@ public class Main {
             //boot error, we should have printed it already
             System.exit(1);
         }
-        modules.add(new FileSystemModule(workingDir, env));
         Injector injector = Guice.createInjector(modules);
         return injector;
     }
@@ -192,9 +215,5 @@ public class Main {
     private static void invalidJarArguments() {
         System.err.println("The first argument of K3 not recognized. Try -kompile, -kast, -krun, -ktest, -kserver, or -kpp.");
         System.exit(1);
-    }
-
-    public static Injector getInjector(String tool, String[] args2) {
-        return getInjector(new File(System.getProperty("user.dir")), System.getenv(), tool, args2);
     }
 }